swap onto our own database class, which is an extended version of PDO.
This currently enables us to track transactions, and to ensure that we set the transaction isolation level to serializable.

Serializable isolation means we effectively hold a write lock on the database while the transaction is pending, though reads are still allowed. This means we have to do any read-before-writes within the transaction to, so we don't risk the data changing in the middle.

As it is a subclass of PDO, I've swapped the system on to the new class by simply changing the constructor call.